Add PointMaker struct
