+++ CoordinateCalculator/CoordinateCalculator/InputChecker.swift
@@ -26,9 +26,7 @@ enum InputError: Error {
 }
 
 struct InputChecker {
-    
-    private let hyphen: Character = "-"
-    private let meaninglessCharacters: CharacterSet = ["(", ")", ","]
+
     static private let validPattern: String = "\\(([1-9]|1[0-9]|2[0-4]),([1-9]|1[0-9]|2[0-4])\\)-?"
     
     // 입력에 좌표문자열패턴이 있는지 검사
@@ -42,37 +40,6 @@ struct InputChecker {
         
         // 입력된 좌표수 체크
         return numberOfMatches < 1 || numberOfMatches > 2
-    }
-    
-    func getMyPoints(_ userInput: String) -> [MyPoint] {
-        let splited: [String] = self.splitByHyphen(in: userInput)
-        let coordinates: [[Int]] = self.getCoordinateFrom(textCoordinates: splited)
-        let myPoints: [MyPoint] = self.makeMyPoint(coordinates)
-        
-        return myPoints
-    }
-    
-    private func splitByHyphen(in text: String) -> [String] {
-        return text.split(separator: hyphen).map(String.init)
-    }
-    
-    private func getCoordinateFrom(textCoordinates: [String]) -> [[Int]] {
-        var coordinates = [[Int]]()
-        for textCoordinate in textCoordinates {
-            coordinates.append(textCoordinate.components(separatedBy: self.meaninglessCharacters).compactMap { Int($0) })
-        }
-        
-        return coordinates
-    }
-    
-    private func makeMyPoint(_ coordinates: [[Int]]) -> [MyPoint] {
-        var myPoints = [MyPoint]()
-        
-        for coordinate in coordinates {
-            myPoints.append(MyPoint(x: coordinate[0], y: coordinate[1]))
-        }
-        
-        return myPoints
     }
     
 }
